backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
transformers==4.36.0
torch==2.1.2
faiss-cpu==1.7.4
pypdf==3.17.4
python-multipart==0.0.6
python-dotenv==1.0.0
tqdm==4.66.1
sentence-transformers==2.2.2
numpy==1.26.2
# Note: BioClinicalBERT model weights will be downloaded from Hugging Face automatically.
